requirements.txt
langchain
langchain[faiss]
deepeval
python-dotenv